+++ common/src/main/java/tocraft/craftedcore/CraftedCoreConfig.java
@@ -2,10 +2,12 @@ package tocraft.craftedcore;
 
 import tocraft.craftedcore.config.Config;
 import tocraft.craftedcore.config.ConfigLoader;
+import tocraft.craftedcore.config.annotions.Comment;
 
 @SuppressWarnings("CanBeFinal")
 public final class CraftedCoreConfig implements Config {
-    public static final CraftedCoreConfig INSTANCE = ConfigLoader.read(CraftedCore.MODID, CraftedCoreConfig.class);
+    public static final CraftedCoreConfig INSTANCE = ConfigLoader.register(CraftedCore.MODID);
+    @Comment("Toggle the version checking for every mod that uses CraftedCore")
     public boolean enableVersionChecking = true;
 
     @Override